Complete 'Use unsigned int for p_max_size in obj_cb.print'

+++ src/audacious.cc
@@ -198,12 +198,12 @@ void print_audacious_status(struct text_object *, char *p, unsigned int p_max_si
 }
 
 void print_audacious_title(struct text_object *obj, char *p, unsigned int p_max_size) {
-  snprintf(p, std::min(obj->data.i, p_max_size), "%s", get_res().title.c_str());
+  snprintf(p, std::min((unsigned int) obj->data.i, p_max_size), "%s", get_res().title.c_str());
 }
 
 void print_audacious_filename(struct text_object *obj, char *p,
                               unsigned int p_max_size) {
-  snprintf(p, std::min(obj->data.i, p_max_size), "%s",
+  snprintf(p, std::min((unsigned int) obj->data.i, p_max_size), "%s",
            get_res().filename.c_str());
 }
